move back requests subscription in devtools page
